2025-09-27/flask_redis_mysql/requirements.txt
flash==1.0.3
Flask==3.1.2
Flask-Limiter==3.12
Flask-MySQL==1.6.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Jinja2==3.1.6
mysql-connector-python==9.4.0
Pygments==2.19.2
PyMySQL==1.1.2
SQLAlchemy==2.0.43
redis==6.4.0


